requirements.txt
appdirs==1.4.3
Django==1.10.6
django-allauth==0.31.0
django-autoslug==1.9.3
django-ckeditor==5.2.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
psycopg2==2.7.1
pyparsing==2.2.0
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
